+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Changed
+
+- Uses of `std::sync::Mutex` and `std::sync::RwLock` changed to their async-aware equivalents from `tokio`.
+- The `Celery::register_task` method is now an async function due to the above.
+
+## Removed
+
+- The `SyncError` variants have been removed.
+
 ## [0.2.0-alpha.0] - 2019-02-17
 
 ### Added